+++ app/Http/Controllers/createproductscontroller.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Properties;
+use App\Property;
 use Illuminate\Http\Request;
-use App\Locations;
-use App\products;
+use App\Location;
+use App\Product;
 use Illuminate\Support\Facades\DB;
 
 class createproductscontroller extends Controller
@@ -13,7 +13,7 @@ class createproductscontroller extends Controller
     public function index()
     {
         return view('products.create', [
-            'locations' => Locations::all()
+            'locations' => Location::all()
         ]);
     }
