+++ app/Location.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Location extends Model
+{
+    public function products()
+    {
+        return $this->hasMany('App\Product');
+    }
+}
