+++ app/Product.php
@@ -4,13 +4,13 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Products extends Model
+class Product extends Model
 {
     /**
      * Get the Property record associated with the product.
      */
     public function properties()
     {
-        return $this->hasOne('App\Properties');
+        return $this->hasOne('App\Property');
     }
 }
