+++ app/Property.php
@@ -4,13 +4,13 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Properties extends Model
+class Property extends Model
 {
     /**
      * Get the Product record associated with the property.
      */
     public function products()
     {
-        return $this->hasOne('App\Products');
+
     }
 }
