--- app/Users.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Users extends Model
-{
-    public function locationuser()
-    {
-        return $this->hasMany('App\Locationuser');
-    }
-}
